E_Commerce/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
